move function to wollmux-core
