--- src/de/muenchen/allg/itd51/wollmux/func/functions/AbsFunction.java
@@ -1,24 +1,0 @@
-package de.muenchen.allg.itd51.wollmux.func.functions;
-
-import java.util.Map;
-
-import de.muenchen.allg.itd51.wollmux.core.dialog.DialogLibrary;
-import de.muenchen.allg.itd51.wollmux.core.functions.FunctionLibrary;
-import de.muenchen.allg.itd51.wollmux.core.parser.ConfigThingy;
-import de.muenchen.allg.itd51.wollmux.core.parser.ConfigurationErrorException;
-
-public class AbsFunction extends SumFunction
-{
-  public AbsFunction(ConfigThingy conf, FunctionLibrary funcLib,
-      DialogLibrary dialogLib, Map<Object, Object> context)
-      throws ConfigurationErrorException
-  {
-    super(conf, funcLib, dialogLib, context);
-  }
-
-  @Override
-  protected String computationResult()
-  {
-    return formatBigDecimal(numericComputationResult().abs());
-  }
-}
