--- src/de/muenchen/allg/itd51/wollmux/func/functions/SignFunction.java
@@ -1,25 +1,0 @@
-package de.muenchen.allg.itd51.wollmux.func.functions;
-
-import java.util.Map;
-
-import de.muenchen.allg.itd51.wollmux.core.dialog.DialogLibrary;
-import de.muenchen.allg.itd51.wollmux.core.functions.FunctionLibrary;
-import de.muenchen.allg.itd51.wollmux.core.parser.ConfigThingy;
-import de.muenchen.allg.itd51.wollmux.core.parser.ConfigurationErrorException;
-
-public class SignFunction extends SumFunction
-{
-  public SignFunction(ConfigThingy conf, FunctionLibrary funcLib,
-      DialogLibrary dialogLib, Map<Object, Object> context)
-      throws ConfigurationErrorException
-  {
-    super(conf, funcLib, dialogLib, context);
-  }
-
-  @Override
-  protected String computationResult()
-  {
-    // signum() liefert int, deswegen hier nur ""+ und nicht formatBigDecimal()
-    return "" + numericComputationResult().signum();
-  }
-}
